requirements.txt
gunicorn
attrs
blinker
certifi
cffi
click
colorama
contourpy 
cycler 
et-xmlfile 
exceptiongroup 
Flask  
fonttools 
h11   
idna  
iniconfig 
itsdangerous 
Jinja2  
kiwisolver  
MarkupSafe 
matplotlib
numpy  
openpyxl  
outcome    
packaging  
pandas       
pathlib  
pillow   
pip    
pluggy    
pycparser  
PyJWT       
pyparsing    
PySocks     
pytest    
python-dateutil   
pytz          
selenium        
setuptools       
six            
sniffio        
sortedcontainers  
tomli            
trio     
trio-websocket  
typing_extensions 
tzdata   
urllib3  
Werkzeug 
wsproto
statsmodels
Pillow
